requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pandas==2.1.3
numpy==1.26.2
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
Pillow==10.1.0
requests==2.31.0
tqdm==4.66.1
faiss-gpu
